+++ backend/src/main/java/toonpick/app/controller/DataRequestController.java
@@ -5,16 +5,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import toonpick.app.dto.AuthorDTO;
-import toonpick.app.dto.GenreDTO;
-import toonpick.app.dto.WebtoonDTO;
-import toonpick.app.dto.WebtoonRequestDTO;
+import toonpick.app.webtoon.dto.AuthorDTO;
+import toonpick.app.webtoon.dto.GenreDTO;
+import toonpick.app.webtoon.dto.WebtoonDTO;
+import toonpick.app.webtoon.dto.WebtoonRequestDTO;
 import toonpick.app.entity.enums.AgeRating;
 import toonpick.app.entity.enums.Platform;
 import toonpick.app.entity.enums.SerializationStatus;
-import toonpick.app.service.AuthorService;
-import toonpick.app.service.GenreService;
-import toonpick.app.service.WebtoonService;
+import toonpick.app.webtoon.service.AuthorService;
+import toonpick.app.webtoon.service.GenreService;
+import toonpick.app.webtoon.service.WebtoonService;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
