+++ backend/src/main/java/toonpick/app/controller/MemberFavoriteWebtoonController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import toonpick.app.auth.user.CustomUserDetails;
-import toonpick.app.dto.WebtoonDTO;
+import toonpick.app.webtoon.dto.WebtoonDTO;
 import toonpick.app.service.MemberFavoriteWebtoonService;
 
 import java.util.List;
