+++ backend/src/main/java/toonpick/app/controller/WebtoonDataUpdateController.java
@@ -7,15 +7,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import toonpick.app.dto.AuthorDTO;
-import toonpick.app.dto.GenreDTO;
-import toonpick.app.dto.WebtoonUpdateRequestDTO;
-import toonpick.app.service.AuthorService;
-import toonpick.app.service.GenreService;
-import toonpick.app.service.WebtoonService;
-
-import java.util.Set;
-import java.util.stream.Collectors;
+import toonpick.app.webtoon.dto.WebtoonUpdateRequestDTO;
+import toonpick.app.webtoon.service.AuthorService;
+import toonpick.app.webtoon.service.GenreService;
+import toonpick.app.webtoon.service.WebtoonService;
 
 @RestController
 @RequestMapping("/api/webtoon-update-request")
