+++ backend/src/main/java/toonpick/app/entity/MemberFavoriteWebtoon.java
@@ -12,6 +12,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import toonpick.app.webtoon.entity.Webtoon;
 
 import java.time.LocalDateTime;
