+++ backend/src/main/java/toonpick/app/repository/ReviewLikeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import toonpick.app.entity.ReviewLike;
 import toonpick.app.entity.Member;
-import toonpick.app.entity.Webtoon;
+import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.entity.WebtoonReview;
 
 import java.util.List;
