+++ backend/src/main/java/toonpick/app/service/WebtoonReviewService.java
@@ -14,13 +14,13 @@ import toonpick.app.dto.WebtoonReviewCreateDTO;
 import toonpick.app.dto.WebtoonReviewDTO;
 import toonpick.app.entity.ReviewLike;
 import toonpick.app.entity.Member;
-import toonpick.app.entity.Webtoon;
+import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.entity.WebtoonReview;
 import toonpick.app.exception.ResourceNotFoundException;
 import toonpick.app.mapper.WebtoonReviewMapper;
 import toonpick.app.repository.ReviewLikeRepository;
 import toonpick.app.repository.MemberRepository;
-import toonpick.app.repository.WebtoonRepository;
+import toonpick.app.webtoon.repository.WebtoonRepository;
 import toonpick.app.repository.WebtoonReviewRepository;
 
 import java.util.List;
