+++ backend/src/main/java/toonpick/app/webtoon/dto/AuthorDTO.java
@@ -1,11 +1,9 @@
-package toonpick.app.dto;
+package toonpick.app.webtoon.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.Set;
 
 @Data
 @NoArgsConstructor
