+++ backend/src/main/java/toonpick/app/webtoon/dto/WebtoonDTO.java
@@ -1,4 +1,4 @@
-package toonpick.app.dto;
+package toonpick.app.webtoon.dto;
 
 import lombok.*;
 import toonpick.app.entity.enums.AgeRating;
