+++ backend/src/main/java/toonpick/app/webtoon/repository/GenreRepository.java
@@ -1,6 +1,6 @@
-package toonpick.app.repository;
+package toonpick.app.webtoon.repository;
 
-import toonpick.app.entity.Genre;
+import toonpick.app.webtoon.entity.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
