+++ backend/src/test/java/toonpick/app/config/TestConfig.java
@@ -3,7 +3,7 @@ package toonpick.app.config;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import toonpick.app.service.AuthorService;
+import toonpick.app.webtoon.service.AuthorService;
 
 @TestConfiguration
 class TestConfig {
